update image pull tests
(cherry picked from commit 46f3d8bb7f73eec26d70853f53d5f5723e3e7a7e)
